Move data blob methods to local participant
